requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.1
click==7.1.2
clickclick==20.10.2
connexion==2.8.0
factory-boy==3.2.0
Faker==8.10.1
Flask==1.1.4
Flask-Caching==1.10.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.12.2
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pymemcache==3.5.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-memcached==1.59
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.20
swagger-ui-bundle==0.0.8
text-unidecode==1.3
urllib3==1.26.6
Werkzeug==1.0.1
